+++ packages/hedron-elements/src/hedron_elements/plugin.py
@@ -38,7 +38,7 @@ _STATIC = _ROOT / "static"
 
 PLUGIN_META = PluginMeta(
     name="hedron_elements",
-    version="1.0.17",
+    version="1.0.18",
     distribution="hedron-elements",
     hedron_version=">=1.0,<2.0",
     capabilities=PluginCapabilities(